# Rewrite the alpha batcher to use a simple shortest-common-subsequence approximation algorithm.