This makes the batcher simpler and more scalable and improves its output
at the cost of a small constant overhead. Because of the scalability
improvements, this patch lifts the cap on the number of alpha batches
per render phase to 4.

Before-and-after measurements on
http://en.wikipedia.org/wiki/Servomechanism:

* Draw calls: 161 → 77 (2.1x fewer)
* Backend CPU time: 3.15ms → 3.32ms (5.4% slower)
* CPU compositor time: 1.03ms → 0.53ms (94% faster)
* GPU time: 1.83ms → 1.80ms (1.7% faster)

Based on #364.

r? @glennw

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/servo/webrender/367)
<!-- Reviewable:end -->