@bors-servo 2016-08-24
:pushpin: Commit d3c1f7c has been approved by `glennw`

<!-- @bors-servo r=glennw d3c1f7c4da617dc6c5e1731658942d407ac37ea0 -->