@bors-servo 2016-08-25
:broken_heart: Test failed - [status-travis](https://travis-ci.org/servo/webrender/builds/154891883)