@bors-servo 2016-08-25
:sunny: Test successful - [status-travis](https://travis-ci.org/servo/webrender/builds/155153147)